refactor: extract compositing into its own module
Closest-hit selection and image assembly were buried in renderer.py
while mesh/render.py carried its own copies. Moving the shape
renderer's versions into composite.py gives the triangle path
something to share when it is rewired.

This is a move only: the ten functions are unchanged and renderer.py
imports back the four it still calls, so every call site is
untouched. The mesh copies stay for now because reconciling two live
implementations that differ in layout, (3, H, W) with a transpose
versus (H, W, 3) directly, would risk the numerics for no gain while
both are in use. They are deleted when the mesh path stops needing
them.

renderer.py drops from 636 to 563 lines.

pytest paz/graphics/: 250 passed, 1 skipped, all twelve goldens
matched.
