+++ paz/graphics/composite.py
@@ -1,0 +1,83 @@
+import jax.numpy as jp
+
+import paz
+
+
+def postprocess(hit_masks, depths, colors, world_to_camera, rays, H, W):
+    hit_masks = jp.expand_dims(hit_masks, 0)
+    depths = jp.expand_dims(depths, 0)
+    colors = jp.expand_dims(colors, 0)
+    scene_hit_mask = compute_scene_hit_mask(hit_masks)
+    scene_colors = select_colors(depths, colors)
+    image = to_color_image(scene_hit_mask, scene_colors, H, W)
+    depth = to_depth_image(scene_hit_mask, depths, world_to_camera, rays, H, W)
+    return image, depth
+
+
+def compute_scene_hit_mask(hit_masks):
+    hit_masks = jp.array(hit_masks)
+    hit_mask = jp.sum(hit_masks, axis=0)
+    hit_mask = hit_mask.astype(bool)
+    return hit_mask
+
+
+def select_colors(depths, colors):
+    depths = jp.array(depths)
+    arg_depths = jp.argmin(depths, axis=0)
+    arg_depths = jp.expand_dims(arg_depths, 0)
+
+    colors = jp.array(colors)
+    ndim = colors.ndim
+    while arg_depths.ndim < ndim:
+        arg_depths = jp.expand_dims(arg_depths, -1)
+
+    colors = jp.take_along_axis(colors, arg_depths, axis=0)
+    colors = jp.squeeze(colors, axis=0)
+    return colors
+
+
+def to_color_image(hit_mask, colors, height, width, background_color=1):
+    image = jp.where(hit_mask, colors.T, background_color)
+    image = image.reshape((3, height, width))
+    image = jp.clip(image[:], 0, 1)
+    image = jp.rollaxis(image, 0, 3)
+    return image
+
+
+def to_depth_image(hit_mask, depths, world_to_camera, rays, height, width):
+    min_depths = jp.min(jp.array(depths), axis=0)
+    world_depths = calculate_world_depth(min_depths, world_to_camera, rays)
+    masked_depths = apply_hit_mask_to_depth(hit_mask, world_depths)
+    return reshape_depth_image(masked_depths, height, width)
+
+
+def calculate_world_depth(min_depths, world_to_camera, rays):
+    points = paz.graphics.geometry.compute_points3D(*rays, min_depths)
+    points = paz.algebra.transform_points(world_to_camera, points)
+    return -points[:, -1]
+
+
+def apply_hit_mask_to_depth(hit_mask, depths, faraway=0):
+    return jp.where(hit_mask, depths, faraway)
+
+
+def reshape_depth_image(depths, height, width):
+    image = depths.reshape((1, height, width))
+    image = jp.rollaxis(image, 0, 3)
+    return image[:, :, 0]
+
+
+def find_closest_intersection_args(hit_masks, depths):
+    if depths.ndim > hit_masks.ndim:
+        depths = jp.squeeze(depths, axis=-1)
+    depths_masked = jp.where(hit_masks, depths, paz.graphics.FARAWAY)
+    closest_args = jp.argmin(depths_masked, axis=0)
+    return closest_args
+
+
+def take_closest(candidates, closest_indices):
+    num_feature_dims = candidates.ndim - 2
+    broadcast_shape = (1, -1) + (1,) * num_feature_dims
+    indices = closest_indices.reshape(broadcast_shape)
+    selected = jp.take_along_axis(candidates, indices, axis=0)
+    return jp.squeeze(selected, axis=0)
